The build of 2.0
Completely changed the whole thing! Now runs on content page, with
a notification center-like thingy to expand and view links, which
persist for a limited time before dissapearing. Can save and hide.
Added a status monitor to check if it's alive, and will restart
the mainloop on errors! Added a whole build system that compiles
and minifies js and css.ø